requirements.txt
# core
flask==3.1.0
Werkzeug==3.1.3
jinja2==3.1.6
WTForms==3.2.1
flask_wtf==1.2.2

# DB
#flask_migrate==4.1.0
#flask_sqlalchemy==3.1.1
#sqlalchemy==2.0.38

google-cloud-datastore

# tools
flask_login==0.6.3
flask-dance==7.1.0
celery==5.4.0
redis==5.2.1
colorama==0.4.6
PyJWT~=2.10.1
#WTForms-Alchemy==0.19.0
Flask-WTF
bootstrap-flask

# utils
email_validator==2.2.0
blinker==1.9.0

# env
python-dotenv==1.0.1

# deployment
gunicorn==23.0.0
Flask-Minify==0.49
Flask-CDN==1.5.3

# flask_mysqldb
# psycopg2-binary
requests
ta
ccxt==4.3.5
pandas